refactor: getAccessToken() 함수 auth로 이동
