+++ backend/src/v1/auth/auth.service.ts
@@ -1,5 +1,6 @@
 import { ResultSetHeader } from 'mysql2';
 import { executeQuery } from '~/mysql';
+import axios from 'axios';
 import { role } from './auth.type';
 
 // eslint-disable-next-line import/prefer-default-export
@@ -14,4 +15,27 @@ export const updateAuthenticationUser = async (
     WHERE id = ?
   `, [intraId, nickname, role.cadet, id]);
   return result.affectedRows;
+};
+
+export const getAccessToken = async (): Promise<string> => {
+  const tokenURL = 'https://api.intra.42.fr/oauth/token';
+  const queryString = {
+    grant_type: 'client_credentials',
+    client_id: process.env.CLIENT_ID,
+    client_secret: process.env.CLIENT_SECRET,
+    redirect_uri: process.env.REDIRECT_URL,
+  };
+  let accessToken: string = '';
+  await axios(tokenURL, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    data: queryString,
+  }).then((response) => {
+    accessToken = response.data.access_token;
+  }).catch((error) => {
+    console.log(error.message);
+  });
+  return accessToken;
 };
