+++ backend/src/v1/books/books.service.ts
@@ -430,29 +430,6 @@ export const updateBookDonator = async (bookDonator: UpdateBookDonator) => {
   await booksRepository.updateBookDonator(bookDonator);
 };
 
-export const getAccessToken = async (): Promise<string> => {
-  const tokenURL = 'https://api.intra.42.fr/oauth/token';
-  const queryString = {
-    grant_type: 'client_credentials',
-    client_id: process.env.CLIENT_ID,
-    client_secret: process.env.CLIENT_SECRET,
-    redirect_uri: process.env.REDIRECT_URL,
-  };
-  let accessToken: string = '';
-  await axios(tokenURL, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    data: queryString,
-  }).then((response) => {
-    accessToken = response.data.access_token;
-  }).catch((error) => {
-    console.log(error.message);
-  });
-  return accessToken;
-};
-
 export const getIntraId = async (
   login: string,
 ): Promise<string> => {
